+++ App/Source/ViewController/MainViewController.swift
@@ -7,8 +7,8 @@ class MainViewController: UITableViewController {
         
         navigationController?.navigationBar.isTranslucent = false
         navigationController?.navigationBar.barStyle = .black
-        navigationController?.navigationBar.tintColor = UIColor.white
-        navigationController?.navigationBar.barTintColor = UIColor(red:0.92, green:0.16, blue:0.19, alpha:1.0)
+        navigationController?.navigationBar.tintColor = .white
+        navigationController?.navigationBar.barTintColor = UIColor(red: 0.92, green: 0.16, blue: 0.19, alpha: 1.0)
         navigationController?.navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor: UIColor.white]
     }