+++ App/Source/ViewController/지도/MapViewController.swift
@@ -10,12 +10,8 @@ class MapViewController: UIViewController, INVMapViewDelegate {
         mapView?.delegate = self
         navigationController?.navigationBar.isTranslucent = false
         navigationController?.navigationBar.barStyle = .black
-        navigationController?.navigationBar.tintColor = UIColor.white
-        navigationController?.navigationBar.barTintColor = UIColor(red:0.92, green:0.16, blue:0.19, alpha:1.0)
+        navigationController?.navigationBar.tintColor = .white
+        navigationController?.navigationBar.barTintColor = UIColor(red: 0.92, green: 0.16, blue: 0.19, alpha: 1.0)
         navigationController?.navigationBar.titleTextAttributes = [NSAttributedString.Key.foregroundColor: UIColor.white]
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
     }
 }